Initial commit
We initiate work on this project to generate TV scripts using RNNs. 
In this commit, we have downloaded the data - scripts from the show 'Seinfeld'. We have cleaned up the data and tokenized the words, including separately tokenizing the punctuation. Next, we batch our data into chunks and use TensorDataset and DataLoader classes to load our data.